old_requirements.txt
openai>=1.40.0
requests>=2.31.0
httpx>=0.27.0
python-dotenv>=1.0.0
pytrends>=4.9.2
supabase>=2.6.0
pydantic>=2.7.0
python-slugify>=8.0.4